#4-1 Datatype 수정

+++ src/main/java/com/core/toy3/src/travel/service/TravelService.java
@@ -24,7 +24,7 @@ public class TravelService {
 
     }
 
-    public Response<TravelResponse> getTravelDetails(Long travelId) {
+    public Response<TravelResponse> getTravelDetails(Integer travelId) {
         return travelRepository.findById(travelId)
                 .map(travel -> Response.response(new TravelResponse(travel.getLikeCount())))
                 .orElse(null);
